moved app into separate file to clarify imports
